requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\gregw\Downloads\rocketWebUI\rocketXbeeClient\client.py: 1
# C:\Users\gregw\Downloads\rocketWebUI\rocketXbeeServer\server.py: 1,2
digi_xbee == 1.4.1

# C:\Users\gregw\Downloads\rocketWebUI\rocketSensorDataCollector\collector.py: 8
pyserial == 3.5

# C:\Users\gregw\Downloads\rocketWebUI\rocketSensorDataCollector\collector.py: 9
smbus == 1.1.post2

# C:\Users\gregw\Downloads\rocketWebUI\rocketChartWebServer\websocket_server.py: 8
tornado == 6.1
